Revert "Provide 2 movement alghoritms (#7)"
This reverts commit a43a81e82b51e3c072b2a4fedde8846aea4389d5.

+++ src/imgui/Point.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Point {
-    int x, y;
-    Point(int x = 0.0f, int y = 0.0f) : x(x), y(y) {}
+    float x, y;
+    Point(float x = 0.0f, float y = 0.0f) : x(x), y(y) {}
 };
